+++ packages/monaco-editor-react/src/indexAllLanguages.tsx
@@ -1,9 +1,0 @@
-export type {
-    MonacoEditorProps
-} from './index.js';
-
-export {
-    MonacoEditorReactComp,
-} from './index.js';
-
-import 'monaco-editor-wrapper/languagesOnly';